Add support for a crash handler
Let user set a command to be run in case of unexpected program
termination; either a crash or the application suddenly exiting for some
other reason.

This enables something similar to what e.g. Plasma does, which runs the
screen locker in the session manager process, so if the screen locker
dies the whole session dies. It's a bit more robust however, as it not
only handles crashes but also exits when we don't expect it
(CVE-2016-2312).