Add option to remove text indicators
Add an additional option -r/--remove-text-indicator to disable
the text indicator when locking or unlocking.
i.e. when it says "Verifying..." or "Wrong!".

This is an option for a minor aesthetic change, but I've made
multiple tweaks such as this in order to customize my desktop
experience and I wonder if others might find this useful too.