declare a development shell in flake.nix
This makes it possible to build i3lock (for development)
on any system on which Nix can be installed (= most Linux systems).

For example, I start Emacs using `nix develop --command emacs`
and that Emacs process is in an environment with all i3lock build deps.

See also:
https://michael.stapelberg.ch/posts/2025-07-27-dev-shells-with-nix-4-quick-examples/

This file is provided best-effort, but PRs are welcome.